rename code generation graph
